requirements.txt
fastapi
uvicorn[standard]
beautifulsoup4
markdown
pyyaml
